Adding memcheck script
This is Professor Thornton's addition to the memcheck function command option in BASH that checks for memory leaks, supposedly
the additional of this script will account for some memory leaks or holes that are more inherent in the C++ language. This should prevent from displaying memory leaks caused by poor user programming.